Refactor command.py to remove unused imports and add placeholder.py for PlaceholderBuild class implementation
